Upstream updates

+++ system/aliases.zsh
@@ -9,23 +9,37 @@ then
   alias la='gls -A --color'
 fi
 
-# Flush the OS X DNS cache
-# @link https://support.apple.com/en-us/HT202516
-alias flush='sudo killall -HUP mDNSResponder'
+## macOS specific aliases
+if test "$(uname)" = "Darwin"; then
 
-# Add a function for Homestead
-# @see https://laravel.com/docs/5.2/homestead#daily-usage
-homestead() {
-	(cd ~/Homestead && vagrant $*)
-}
+    ## Flush the DNS cache Mac OS X 10.7+
+	## @link https://support.apple.com/en-us/HT202516
+    alias flushdns='sudo killall -HUP mDNSResponder'
 
-# Launch homestead
-alias hu='homestead up'
-alias hh='homestead halt'
+    alias copy='pbcopy'
+    alias paste='pbpaste'
+    alias desktop="cd $HOME/Desktop"
+    alias download="cd $HOME/Downloads"
+fi
 
-# Launch homestead and ssh into the box
-#alias work='homestead up && homestead ssh -- -t "cd Sites; /usr/bin/zsh"'
-alias work='~/bin/work'
+## Laravel Homestead specific aliases
+if [ -d "~/Homestead" ]; then
+
+	# Add a function for Homestead
+	# @see https://laravel.com/docs/5.8/homestead#daily-usage
+	homestead() {
+		(cd ~/Homestead && vagrant $*)
+	}
+
+
+	# Launch homestead
+	alias hu='homestead up'
+	alias hh='homestead halt'
+
+	# Launch homestead and ssh into the box
+	#alias work='homestead up && homestead ssh -- -t "cd Sites; /usr/bin/zsh"'
+	alias work='~/bin/work'
+fi
 
 # An alias for the Behat testing framework
 alias behat='vendor/bin/behat'
@@ -44,3 +58,18 @@ alias al='php artisan list'
 composer-link() {  
     composer config repositories.local '{"type": "path", "url": "'$1'"}' --file composer.json
 }
+
+
+
+
+
+
+## Laravel aliases
+alias pa='php artisan'
+alias pal='php artisan list'
+
+alias yrd='yarn run dev'
+alias yw='yarn run watch'
+
+# https://superuser.com/a/975878/9551
+alias brewski='brew update && brew upgrade && brew cleanup; brew doctor'
